guqin_jzp/JZPNet/requirements.txt
editdistance==0.5.3
einops==0.3.0
numpy==1.18.5
pandas==1.1.4
Pillow==9.1.0
pytorch_lightning==1.5.8
torch==1.10.1
torchmetrics==0.6.2
torchvision==0.11.2
tqdm==4.62.3
